+++ Sources/BitcoinKit/Core/Mnemonic/Mnemonic.swift
@@ -103,7 +103,7 @@ public struct Mnemonic {
         var bytes = Data(count: byteCount)
         let status = bytes.withUnsafeMutableBytes { SecRandomCopyBytes(kSecRandomDefault, byteCount, $0) }
         guard status == errSecSuccess else { throw MnemonicError.randomBytesError }
-        return generate(entropy: bytes, language: language)
+        return try generate(entropy: bytes, language: language)
     }
 
     public static func deriveLanguageFromMnemonic(words: [String]) -> Language? {
@@ -127,53 +127,48 @@ public struct Mnemonic {
         return nil
     }
 
-    public static func validateChecksumDerivingLanguageOf(mnemonic mnemonicWords: [String]) throws {
+    @discardableResult
+    public static func validateChecksumDerivingLanguageOf(mnemonic mnemonicWords: [String]) throws -> Bool {
         guard let derivedLanguage = deriveLanguageFromMnemonic(words: mnemonicWords) else {
             throw MnemonicError.invalid(.unableToDeriveLanguageFrom(words: mnemonicWords))
         }
-        try validateChecksumOf(mnemonic: mnemonicWords, language: derivedLanguage)
+        return try validateChecksumOf(mnemonic: mnemonicWords, language: derivedLanguage)
     }
 
     // https://github.com/mcdallas/cryptotools/blob/master/btctools/HD/__init__.py#L27-L41
     // alternative in C:
     // https://github.com/trezor/trezor-crypto/blob/0c622d62e1f1e052c2292d39093222ce358ca7b0/bip39.c#L161-L179
-    public static func validateChecksumOf(mnemonic mnemonicWords: [String], language: Language) throws {
+    @discardableResult
+    public static func validateChecksumOf(mnemonic mnemonicWords: [String], language: Language) throws -> Bool {
         let vocabulary = wordList(for: language)
 
-//        let indexes: [UInt11] = try mnemonicWords.map { word in
-//            guard let indexInVocabulary = vocabulary.firstIndex(of: word) else {
-//                throw MnemonicError.invalid(.wordNotInList(word, language: language))
-//            }
-//            guard let indexAs11Bits = UInt11(exactly: indexInVocabulary) else {
-//                fatalError("Unexpected error, is word list longer than 2048 words, it shold not be")
-//            }
-//            return indexAs11Bits
-//        }
+        let indices: [UInt11] = try mnemonicWords.map { word in
+            guard let indexInVocabulary = vocabulary.firstIndex(of: word) else {
+                throw MnemonicError.invalid(.wordNotInList(word, language: language))
+            }
+            guard let indexAs11Bits = UInt11(exactly: indexInVocabulary) else {
+                fatalError("Unexpected error, is word list longer than 2048 words, it shold not be")
+            }
+            return indexAs11Bits
+        }
 
-        let bitStringFromIndex: String = try mnemonicWords.map { word in
-             guard let indexInVocabulary = vocabulary.firstIndex(of: word) else {
-                 throw MnemonicError.invalid(.wordNotInList(word, language: language))
-             }
-            let binString = String(indexInVocabulary, radix: 2).prepending(element: "0", toLength: 11)
-            assert(binString.count == 11)
-            return binString
-        }.joined()
+        let bitArray = BitArray(indices)
 
-        assert(bitStringFromIndex.count == mnemonicWords.count * 11)
-        let bitArray = BitArray(binaryString: bitStringFromIndex)!
         let checksumLength = mnemonicWords.count / 3
+
         let dataBits = bitArray.prefix(subtractFromCount: checksumLength)
         let checksumBits = bitArray.suffix(maxBitCount: checksumLength)
-        let data = dataBits.asData()
-        let hash = Crypto.sha256(data)
-        let hashBinary = BitArray(data: hash)
-        let hashPadded = hashBinary.appending(element: false, toLength: 256)
-        let hashBits = hashPadded.prefix(maxBitCount: checksumLength)
+
+        let hash = Crypto.sha256(dataBits.asData())
+
+        let hashBits = BitArray(data: hash).prefix(maxBitCount: checksumLength)
+
         guard hashBits == checksumBits else {
             throw MnemonicError.invalid(.checksumMismatch)
         }
 
         // All is well
+        return true
     }
 
     private static func intToBinString<I>(_ int: I) -> String where I: BinaryInteger {
@@ -185,8 +180,12 @@ public struct Mnemonic {
         return String(("00000000" + String(byte, radix: 2)).suffix(8))
       }
 
-    internal static func generate(entropy: Data, language: Language = .english) -> [String] {
-       let entropybits = String(entropy.flatMap { ("00000000" + String($0, radix: 2)).suffix(bitsPerByte) })
+    internal static func generate(
+        entropy: Data,
+        language: Language = .english
+    ) throws -> [String] {
+
+        let entropybits = String(entropy.flatMap { ("00000000" + String($0, radix: 2)).suffix(bitsPerByte) })
         let hash = Crypto.sha256(entropy)
         let hashBits = String(hash.flatMap { ("00000000" + String($0, radix: 2)).suffix(bitsPerByte) })
         let checkSum = String(hashBits.prefix((entropy.count * bitsPerByte) / 32))
@@ -202,11 +201,20 @@ public struct Mnemonic {
             mnemonic.append(String(words[wordIndex]))
         }
 
+        try validateChecksumOf(mnemonic: mnemonic, language: language)
+
         return mnemonic
     }
 
-    public static func seed(mnemonic words: [String], passphrase: String = "") throws -> Data {
-        try validateChecksumDerivingLanguageOf(mnemonic: words)
+    /// Pass a trivial closure: `{ _ in }` to `validateChecksum` if you would like to opt-out of checksum validation.
+    public static func seed(
+        mnemonic words: [String],
+        passphrase: String = "",
+        validateChecksum: (([String]) throws -> Void) = { try Mnemonic.validateChecksumDerivingLanguageOf(mnemonic: $0) }
+    ) rethrows -> Data {
+
+        try validateChecksum(words)
+
         let mnemonic = words.joined(separator: " ").decomposedStringWithCompatibilityMapping.data(using: .utf8)!
         let salt = ("mnemonic" + passphrase).decomposedStringWithCompatibilityMapping.data(using: .utf8)!
         let seed = _Key.deriveKey(mnemonic, salt: salt, iterations: 2048, keyLength: 64)
@@ -235,92 +243,6 @@ public struct Mnemonic {
     }
 }
 
-struct UInt11: ExpressibleByIntegerLiteral {
-    static var bitWidth: Int { 11 }
-
-    static var max16: UInt16 { UInt16(2047) }
-    static var max: UInt11 { UInt11(exactly: max16)! }
-
-    static var min: UInt11 { 0 }
-
-    private let valueBoundBy16Bits: UInt16
-
-    init?<T>(exactly source: T) where T: BinaryInteger {
-        guard
-            let valueBoundBy16Bits = UInt16(exactly: source),
-            valueBoundBy16Bits < 2048 else { return nil }
-
-        self.valueBoundBy16Bits = valueBoundBy16Bits
-    }
-
-}
-
-extension UInt11 {
-    init<T>(truncatingIfNeeded source: T) where T: BinaryInteger {
-         let valueBoundBy16Bits = UInt16(truncatingIfNeeded: source)
-         self.valueBoundBy16Bits = Swift.min(UInt11.max16, valueBoundBy16Bits)
-     }
-
-     /// Creates a new integer value from the given string and radix.
-     init?<S>(_ text: S, radix: Int = 10) where S: StringProtocol {
-         guard let uint16 = UInt16(text, radix: radix) else { return nil }
-         self.init(exactly: uint16)
-     }
-
-    init(integerLiteral value: Int) {
-        guard let exactly = UInt11(exactly: value) else {
-            fatalError("bad integer literal value does not fit in UInt11, value passed was: \(value)")
-        }
-        self = exactly
-    }
-}
-
-extension UInt11 {
-    var binaryString: String {
-        let binaryString = String(valueBoundBy16Bits.binaryString.suffix(Self.bitWidth))
-        assert(UInt16(binaryString, radix: 2)! == valueBoundBy16Bits, "expected '\(UInt16(binaryString, radix: 2)!)' to equal: '\(valueBoundBy16Bits)', bound.valueBoundBy16Bits.binaryString: \(valueBoundBy16Bits.binaryString), self.binaryString: \(binaryString)")
-        return binaryString
-
-//        var result: [String] = []
-//        for i in 0..<(Self.bitWidth / 8) {
-//            let byte = UInt8(truncatingIfNeeded: self >> (i * 8))
-//            let byteString = String(byte, radix: 2)
-//            let padding = String(repeating: "0",
-//                                 count: 8 - byteString.count)
-//            result.append(padding + byteString)
-//        }
-//        return "0b" + result.reversed().joined(separator: "_")
-    }
-}
-
-extension FixedWidthInteger {
-    var binaryString: String {
-        var result: [String] = []
-        for i in 0..<(Self.bitWidth / 8) {
-            let byte = UInt8(truncatingIfNeeded: self >> (i * 8))
-            let byteString = String(byte, radix: 2)
-            let padding = String(repeating: "0",
-                                 count: 8 - byteString.count)
-            result.append(padding + byteString)
-        }
-        return result.reversed().joined()
-    }
-}
-
-extension Data {
-    var binaryString: String {
-        var result: [String] = []
-        for byte in self {
-//            let byte = UInt8(truncatingIfNeeded: self >> (i * 8))
-            let byteString = String(byte, radix: 2)
-            let padding = String(repeating: "0",
-                                 count: 8 - byteString.count)
-            result.append(padding + byteString)
-        }
-        return result.joined()
-    }
-}
-
 public enum MnemonicError: Error {
     case randomBytesError
 
@@ -331,43 +253,4 @@ public enum MnemonicError: Error {
         case unableToDeriveLanguageFrom(words: [String])
         case checksumMismatch
     }
-}
-private extension BinaryInteger {
-    var asData: Data {
-        var int = self
-        return Data(bytes: &int, count: MemoryLayout<Self>.size)
-    }
-}
-
-extension RangeReplaceableCollection {
-
-    mutating func prepend(element: Element, toLength expectedLength: Int?) {
-        self = prepending(element: element, toLength: expectedLength)
-    }
-
-    func prepending(element: Element, toLength expectedLength: Int?) -> Self {
-        guard let expectedLength = expectedLength else {
-            return self
-        }
-        var modified = self
-        while modified.count < expectedLength {
-            modified = [element] + modified
-        }
-        return modified
-    }
-
-    mutating func append(element: Element, toLength expectedLength: Int?) {
-          self = appending(element: element, toLength: expectedLength)
-      }
-
-      func appending(element: Element, toLength expectedLength: Int?) -> Self {
-          guard let expectedLength = expectedLength else {
-              return self
-          }
-          var modified = self
-          while modified.count < expectedLength {
-            modified.append(element)
-          }
-          return modified
-      }
 }
